x86/svm: Track and handle exit code as an unsigned 64-bit value
Track and handle SVM's exit_code as an unsigned 64-bit value.  Per the
APM, offset 0x70 is a single 64-bit value:

  070h 63:0 EXITCODE

And a sane reading of the error values defined in "Table C-1. SVM Intercept
Codes" is that negative values use the full 64 bits:

  –1 VMEXIT_INVALID Invalid guest state in VMCB.
  –2 VMEXIT_BUSYBUSY bit was set in the VMSA
  –3 VMEXIT_IDLE_REQUIREDThe sibling thread is not in an idle state
  -4 VMEXIT_INVALID_PMC Invalid PMC state

And that interpretation is confirmed by testing on Milan and Turin (by
setting bits in CR0[63:32] to generate VMEXIT_INVALID on VMRUN).

Furthermore, Xen has treated exitcode as a 64-bit value since HVM support
was adding in 2006 (see Xen commit d1bd157fbc ("Big merge the HVM
full-virtualisation abstractions.")).

Note, the SVM tests will fail when on KVM builds without commit
f402ecd7a8b6 ("KVM: nSVM: Set exit_code_hi to -1 when synthesizing
SVM_EXIT_ERR (failed VMRUN)").

Link: https://lore.kernel.org/all/20251113225621.1688428-1-seanjc@google.com
Link: https://patch.msgid.link/20251230191342.4052363-1-seanjc@google.com

+++ x86/svm.c
@@ -244,7 +244,7 @@ void svm_setup_vmrun(u64 rip)
 	vmcb->save.rsp = (ulong)(guest_stack + ARRAY_SIZE(guest_stack));
 }
 
-int __svm_vmrun(u64 rip)
+u64 __svm_vmrun(u64 rip)
 {
 	svm_setup_vmrun(rip);
 	regs.rdi = (ulong)v2_test;
@@ -260,7 +260,7 @@ int __svm_vmrun(u64 rip)
 	return (vmcb->control.exit_code);
 }
 
-int svm_vmrun(void)
+u64 svm_vmrun(void)
 {
 	return __svm_vmrun((u64)test_thunk);
 }
